core_engine/requirements.txt
psycopg2-binary==2.8.5
aiofiles==0.5.0
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
fastapi==0.55.1
h11==0.9.0
httptools==0.0.13
idna==2.8
numpy==1.18.4
pandas==1.0.3
prompt-toolkit==2.0.10
psycopg2-binary==2.8.5
pydantic==1.4
Pygments==2.3.1
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2020.1
PyYAML==5.3.1
requests==2.22.0
six==1.15.0
starlette==0.13.2
urllib3==1.24.3
uvicorn==0.11.2
uvloop==0.14.0
wcwidth==0.1.9
websockets==8.1
boto3==1.15.2
google-cloud-automl==2.4.0
google-cloud-storage==1.41.1